Move some packages to api module
