+++ api/src/main/java/me/saiintbrisson/minecraft/event/EventSubscription.java
