+++ api/src/main/java/me/saiintbrisson/minecraft/internal/ContainerModification.java
