+++ api/src/main/java/me/saiintbrisson/minecraft/pipeline/PipelineInterceptor.java
