+++ shared/src/main/java/me/saiintbrisson/minecraft/pipeline/interceptors/ScheduledUpdateInterceptor.java
@@ -4,8 +4,8 @@ import java.time.Duration;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
-import me.saiintbrisson.minecraft.Job;
 import me.saiintbrisson.minecraft.ViewContext;
+import me.saiintbrisson.minecraft.internal.Job;
 import me.saiintbrisson.minecraft.pipeline.PipelineContext;
 import me.saiintbrisson.minecraft.pipeline.PipelineInterceptor;
 import org.jetbrains.annotations.NotNull;
