+++ shared/src/test/java/me/saiintbrisson/minecraft/pipeline/interceptors/ScheduledUpdateInterceptorTest.java
@@ -12,9 +12,9 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.concurrent.atomic.AtomicBoolean;
 import me.saiintbrisson.minecraft.AbstractView;
-import me.saiintbrisson.minecraft.Job;
 import me.saiintbrisson.minecraft.ViewContext;
 import me.saiintbrisson.minecraft.Viewer;
+import me.saiintbrisson.minecraft.internal.Job;
 import me.saiintbrisson.minecraft.pipeline.Pipeline;
 import org.junit.jupiter.api.Test;
