+++ test/src/main/java/me/saiintbrisson/minecraft/test/TestViewFrame.java
@@ -8,7 +8,6 @@ import java.util.function.UnaryOperator;
 import lombok.Getter;
 import lombok.Setter;
 import me.saiintbrisson.minecraft.AbstractView;
-import me.saiintbrisson.minecraft.Job;
 import me.saiintbrisson.minecraft.PaginatedViewContext;
 import me.saiintbrisson.minecraft.PlatformViewFrame;
 import me.saiintbrisson.minecraft.ViewComponentFactory;
@@ -16,6 +15,7 @@ import me.saiintbrisson.minecraft.ViewErrorHandler;
 import me.saiintbrisson.minecraft.ViewItem;
 import me.saiintbrisson.minecraft.Viewer;
 import me.saiintbrisson.minecraft.feature.Feature;
+import me.saiintbrisson.minecraft.internal.Job;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
